refactor(#2679): extract store_url module for QUAL-10 ceiling
CI RED root cause: src/daemon_runtime.rs at 11_934 LOC > QUAL-10 ceiling
11_860 (tests/qual_10_module_size_ceiling.rs). Unit + binary #2679 tests
already passed; only the module-size gate failed.

Extract the store-URL surface into src/store_url.rs (scheme sniff #2444,
non-argv channels #1927, fail-closed refuse #2679 + unit tests). Re-export
from daemon_runtime so migrate / cli::backup / historical paths stay
stable. Pub(crate) store_url_env_lock remains the single process-global
mutex so fx_f2_build_store_handle_no_url cannot race siblings.

Measured: daemon_runtime 11_607 (ceiling 11_660), store_url 361 (ceiling
403). QUAL-10, issue_1927_*, issue_2679_*, store_url_fail_closed_2679, and
clippy pedantic all green.

Rust skills: proj-mod-by-feature, M-BALANCED-MODULES, own-borrow-over-clone,
err-result-over-panic, obs-no-sensitive-data (redact on refuse path).
