+++ tests/qual_10_module_size_ceiling.rs
@@ -1393,7 +1393,8 @@ const MODULE_SIZE_CEILINGS: &[(&str, usize)] = &[
     // coverage tests (bind/revoke × json/non-json + empty-token error) to clear
     // the daemon_runtime.rs per-module COVERAGE floor. Combined with #2045 L6
     // this lands daemon_runtime.rs at 10_676; ceiling 10_500 → 10_750.
-    ("src/daemon_runtime.rs", 11_860), /* 2026-07-31 #2579/#2583 — two paced background loops are wired into `bootstrap_serve` (the FTS5 integrity checker whose cached verdict `/health` renders, and the corpus-size gauge refresher that takes `db::stats` off the /metrics scrape path), plus the sqlite-backend gate on each. The gate is a CORRECTNESS condition, not an optimisation — `db_path`/`db_state` name the local sqlite file, which on a `--store-url postgres://…` daemon is a near-empty sidecar, so an ungated checker would assert integrity over a database nobody reads (the #2444 shape). The spawn-count pin in `test_bootstrap_serve_keyword_tier_no_embedder` moves 8 -> 10 in lockstep and its comment now enumerates the members instead of restating a total that had already drifted (it read "Nine" while asserting 8). Lands the file at 11_793; ceiling 11_780 -> 11_860 (+67 headroom). */
+    ("src/daemon_runtime.rs", 11_660), /* 2026-08-03 #2679/#2680 QUAL-10: extracted store-URL helpers + unit tests to `src/store_url.rs`. Measured 11610; ceiling 11_860 -> 11_660 (+50 headroom). PRIOR: 2026-07-31 #2579/#2583 — two paced background loops are wired into `bootstrap_serve` (the FTS5 integrity checker whose cached verdict `/health` renders, and the corpus-size gauge refresher that takes `db::stats` off the /metrics scrape path), plus the sqlite-backend gate on each. The gate is a CORRECTNESS condition, not an optimisation — `db_path`/`db_state` name the local sqlite file, which on a `--store-url postgres://…` daemon is a near-empty sidecar, so an ungated checker would assert integrity over a database nobody reads (the #2444 shape). The spawn-count pin in `test_bootstrap_serve_keyword_tier_no_embedder` moves 8 -> 10 in lockstep and its comment now enumerates the members instead of restating a total that had already drifted (it read "Nine" while asserting 8). Lands the file at 11_793; ceiling 11_780 -> 11_860 (+67 headroom). */
+    ("src/store_url.rs", 403), /* 2026-08-03 #2679/#2680 QUAL-10: new module extracted from daemon_runtime (scheme sniff + #1927 channels + #2679 fail-closed refuse + unit tests). Measured 363; ceiling 403 (+40 headroom). */
     /* 2026-07-31 #2490 (follow-up) `test_run_dispatch_export_command` asserted the phantom-DB defect as expected behaviour (it dispatched `export` at a TestEnv path where no database was ever created and unwrapped the success); it is split into a refusal cell asserting the REASON plus a CONTROL proving a real corpus still exports with exit 0 and a non-empty artifact, landing the file at 11_719; ceiling 11_700 -> 11_780 (+61 headroom, lockstep). 2026-07-31 #2490 export/import return an exit code so a PARTIAL export or an incomplete import cannot present as success; the two dispatch arms grow from a bare tail-call to the `match ... { 0 => Ok(()), code => exit(code) }` shape the `Command::Config` precedent already uses, landing the file at 11_647; ceiling 11_640 -> 11_700 (+53 headroom, lockstep). */
     /* 2026-07-17 #2167 extract run_sqlite_embedding_space_boot_maintenance helper + both-open-arms unit test to cover the boot-open Err arm (10_783). 2026-07-19 #2064 erasure gc-tick wiring stacked on the #2205 export --full + #1860 vectorlite serve/mcp boot funnels lands the merged file at 10_853; ceiling 10_850 -> 10_900 (lockstep). 2026-07-19 pre-ship 3x7: erasure sweep moved OFF the handler mutex (detached spawn_blocking arm + shared log helper) lands the file at 10_931; ceiling 10_900 -> 10_950 (lockstep). 2026-07-19 merged with the #2233/#2235 lineage boot-seed (+26) -> 10_957; ceiling 10_950 -> 11_000 (lockstep). 2026-07-19 #2271 consultation-posture mutation seam + behavior pin lands at 11_032; ceiling 11_000 -> 11_050 (lockstep). 2026-07-20 #2271 shutdown lifecycle hardening tracks every writer, bounds plain/TLS quiescence, drains deferred audit, and makes final witness/WAL certification fail closed; production plus regression coverage lands at 11_326, ceiling 11_050 -> 11_400 (+74 headroom, lockstep). 2026-07-21 #2290 sign the sync_cycle_once /sync/since pull GET (load daemon signing key + attach X-Memory-Sig/X-Memory-Nonce) lands the file at 11_406; ceiling 11_400 -> 11_450 (+44 headroom, lockstep). 2026-07-23 FBL-22 postgres serve maintenance loop (spawn_postgres_maintenance_loop_if_enabled gc + archive-purge + lease-sweep pg twin + bootstrap wiring + 2 spawn/skip unit tests) lands the file at 11_573; ceiling 11_450 -> 11_640 (+67 headroom, lockstep). */
     ("src/subscriptions.rs", 4_520), /* 2026-07-31 #2445 — 2-line disposition note on the ordering-covered raw open; measured 4_502 */
